backend/pyproject.toml
[tool.poetry]
name = "meeting-summarizer-backend"
version = "0.1.0"
description = "Backend for real-time meeting summarization with local LLM inference"
authors = ["Your Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
llama-cpp-python = "^0.2.0"
tokenizers = "^0.15.0"
redis = "^5.0.1"
python-multipart = "^0.0.6"
websockets = "^12.0"
httpx = "^0.25.2"
python-json-logger = "^2.0.7"
prometheus-client = "^0.19.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
